Subnet & route table creation
For each element in the 'public_subnet_cidrs' list a public subnet is
created and for each element in the 'private_subnet_cidrs' a private
subnet. The public/private subnet are put inside the different
'availibility_zones'. Anymore two route tables are created, one for all
public subnets and one for all private subnets.